+++ src/GeoEfficiency.jl
@@ -35,6 +35,7 @@ about() = print_with_color(:green,"""\n
 export about,
 		
     # Input_Interface
+	SetSrcToPoint,
 	getfloat, 
 	getDetectors,