+++ News.md
@@ -2,6 +2,7 @@
 
 ### Version 0.8.7-Dev
 * new function about() give information about the sofftware Package.
+* new function SetSrcToPoint() to set source type.
 
 ### Version 0.8.6
 * When batch taking arrguments. Any of the arrays `srcHeights_array`, `srcRhos_array`, `srcRadii_array`, `srcLengths_array` element type should be float64. If any of them have Real element type it should converted float64 to using `float` befor passing to the `batch` function.